Cleanup
* Use configured keypair path
* Make num_users configurable
* Make testnet benchmark an example, not a test
