+++ themis/client/Cargo.toml
@@ -20,21 +20,20 @@ bincode = "1.3"
 borsh = "0.7.1"
 bn = {git = "https://github.com/garious/bn", rev = "5c35c737ffabac9921310f53f48725216d59cbf1", default-features = false, features = ["borsh"]}
 elgamal_bn = { git = "https://github.com/garious/elgamal_bn", rev = "ba9bdcdb6cdd6fb8e74d0b8bc1b918bcd1b543a9" }
-#elgamal_bn = { path = "../../../elgamal_bn" }
 futures = "0.3"
 solana-banks-client = "1.3.14"
+solana-cli-config = "1.3.14"
 solana-sdk = "1.3.14"
 spl-themis = { version = "0.1.0", path = "../program" }
 tarpc = { version = "0.21.1", features = ["full"] }
+tokio = "0.2"
+url = "2.1"
 
 [dev-dependencies]
-tokio = "0.2"
 solana-banks-server = "1.3.14"
 solana-bpf-loader-program = "1.3.14"
-solana-cli-config = "1.3.14"
 solana_rbpf = "=0.1.31"
 solana-runtime = "1.3.14"
-url = "2.1"
 
 [lib]
 crate-type = ["cdylib", "lib"]
