+++ themis/program/Cargo.toml
@@ -22,7 +22,6 @@ bincode = "1.3"
 borsh = "0.7.1"
 bn = {git = "https://github.com/garious/bn", rev = "5c35c737ffabac9921310f53f48725216d59cbf1", default-features = false, features = ["borsh"]}
 elgamal_bn = { git = "https://github.com/garious/elgamal_bn", rev = "ba9bdcdb6cdd6fb8e74d0b8bc1b918bcd1b543a9" }
-#elgamal_bn = { path = "../../../elgamal_bn" }
 getrandom = { version = "0.1.14", features = ["dummy"] }
 num-derive = "0.3"
 num-traits = "0.2"
